+++ packages/core-js/features/map/delete-all.js
@@ -1,0 +1,4 @@
+require('../../modules/es.map');
+require('../../modules/esnext.map.delete-all');
+
+module.exports = require('../../internals/entry-unbind')('Map', 'deleteAll');