+++ packages/core-js/features/weak-map/delete-all.js
@@ -1,0 +1,4 @@
+require('../../modules/es.weak-map');
+require('../../modules/esnext.weak-map.delete-all');
+
+module.exports = require('../../internals/entry-unbind')('WeakMap', 'deleteAll');