+++ packages/core-js/internals/collection-delete-all.js
@@ -4,11 +4,11 @@ var aFunction = require('../internals/a-function');
 
 // https://github.com/tc39/collection-methods
 module.exports = function (/* ...elements */) {
-  var set = anObject(this);
-  var remover = aFunction(set['delete']);
+  var collection = anObject(this);
+  var remover = aFunction(collection['delete']);
   var result = true;
   for (var k = 0, len = arguments.length; k < len; k++) {
-    if (!remover.call(set, arguments[k])) result = false;
+    if (!remover.call(collection, arguments[k])) result = false;
   }
   return result;
 };